chore: Rename files using kebab casing
